+++ src/VarDump/Utils/ReflectionUtils.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text;
-using VarDump.CodeDom;
+using VarDump.CodeDom.CSharp;
 using VarDump.CodeDom.VisualBasic;
 using VarDump.Extensions;
